Projet/requirements.txt
flask
flasgger
pandas
numpy
scikit-learn
mlflow
matplotlib
pytest
statsmodels
PyJWT
pytest
requests
streamlit
sqlalchemy
PyYAML
opencv-python
plotly
pymongo
beautifulsoup4
ultralytics
atoti
psutil
docker
cloudpickle